artifact/requirements.txt
opencv-python
imread-from-url
onnxruntime
cap-from-youtube
awsiotsdk
torch